use longs instead of ints to store pointers in NIO buffers
This allows the code to be used on 64-bit VMs.

+++ luni/src/main/java/java/nio/NioUtils.java
@@ -41,7 +41,7 @@ public final class NioUtils {
      * @return the address of the buffer given, or zero if the buffer is not a
      *         direct Buffer.
      */
-    public static int getDirectBufferAddress(Buffer buffer) {
+    public static long getDirectBufferAddress(Buffer buffer) {
         return buffer.effectiveDirectAddress;
     }
